refactor: rename functions according to task description
