+++ product-service/src/functions/get-products-by-id/index.ts
