+++ product-service/src/functions/get-products-by-id/schema.ts
