+++ product-service/src/functions/get-products-list/index.ts
