+++ product-service/src/functions/get-products-list/schema.ts
