+++ product-service/src/functions/index.ts
@@ -1,2 +1,2 @@
-export { default as getProducts } from './get-products';
-export { default as getProductById } from './get-product-by-id';
+export { default as getProductsList } from './get-products-list';
+export { default as getProductsById } from './get-products-by-id';
